Prevent error from using "Bring Forward" on last element.
If "Bring Forward" is clicked on the last element in the DOM tree, the console displayed an error, as it cannot set the attribute of "undefined" which is a result of looking for an element at an index "+1" greater than the last index.

This update prevents updateZIndex from running "Bring Forward" if this function is used on the last element.